編集：User::IllustratedBooksController showアクションのoptionsを削除
編集：LikesController
createアクションにdestroy処理を集約
アイコンの状態を操作するためshowアクションにてlike済みか否かの判定

+++ app/controllers/api/v1/likes_controller.rb
@@ -1,14 +1,25 @@
 class Api::V1::LikesController < ApplicationController
 
-  def create
-    like = current_user.likes.create(illustrated_book_id: params[:illustrated_book_id])
-    render json: like
+  def show
+    illustrated_book_id = params[:id]
+    like = current_user.likes.find_by(illustrated_book_id: illustrated_book_id)
+    if like
+      render json: {likeExists: true}
+    else
+      render json: {likeExists: false}
+    end
   end
 
-  def destroy
-    illustrated_book = IllustratedBook.find(params[:illustrated_book_id])
-    like = current_user.likes.find_by(illustrated_book_id: illustrated_book.id)
-    like.destroy
-    render json
+  def create
+    illustrated_book_id = params[:illustrated_book_id]
+    like = current_user.likes.find_by(illustrated_book_id: illustrated_book_id)
+    if like
+      like.destroy
+      render json: {status: 'ok'}
+    else
+      like = current_user.likes.create(illustrated_book_id: illustrated_book_id)
+      json_string = IllustratedBookSerializer.new(like.illustrated_book).serializable_hash.to_json
+      render json: json_string
+    end
   end
 end
